Correcting some issue with the datasets:
BNCI2015001: correcting the interval. This was done by first trying to checking the description which has no indication whether stim is the start of the trial or appearance of the cue. Therefore, I tested both on CrossSubjectEvaluation using CSP+LDA and MDRM as classifier and the one with higher results was chosen. The chosen update showed a roc_auc increase of 0.15, 0.13 for CSP+LDA and MDRM, respectively.

MunichMI: The annotation and interval were chosen based on the description of the dataset (i.e. the interval between chosen annotation and the next should be 7s).

Ofner2017: was dealt with similar to BNCI2015004 (with right elbow flexion and rest as the algorithms used can't distish between right elbow flexion and right elbow extension).

Zhou2016: was dealt with similar to BNCI2015004. However, unlike the other the performance improvements were minimum as there are a big overlap between them. Better performance gains can be seen with using the 3 classes.

Overall, all datasets either riley on mne.find_events(.) or mne.event_from_annotations with providing event_id.